+++ src/components/Container/Container.js
@@ -1,0 +1,12 @@
+import cn from 'classnames';
+import s from './style.module.css';
+
+const Container = ({children, className}) => {
+    return (
+        <div className={cn(s.root, className)}>
+            {children}
+        </div>
+    );
+}
+ 
+export default Container;
