+++ src/components/Container/style.module.css
@@ -1,0 +1,26 @@
+.root {
+  position: relative;
+  margin: 0 50px;
+  color: #444;
+  max-width: 1180px;
+  width: 100%;
+  overflow: hidden;
+}
+
+@media (max-width: 1024px) {
+  .root {
+    margin: 0 25px;
+  }
+}
+
+@media (max-width: 768px) {
+  .root {
+    margin: 0 14px;
+  }
+}
+
+@media (max-width: 560px) {
+  .root {
+    margin: 0 8px;
+  }
+}
