+++ src/components/Footer/Footer.js
@@ -1,0 +1,18 @@
+import Container from "../Container";
+import s from "./style.module.css";
+
+const Footer = () => {
+  return (
+    <footer className={s.root}>
+      <Container>
+        <div className={s.footerWrap}>
+          Coded with
+          <span className={s.heart}></span>
+          by You
+        </div>
+        </Container>
+    </footer>
+  );
+};
+
+export default Footer;
