+++ src/components/Footer/index.js
@@ -1,17 +1,3 @@
-import s from "./style.module.css";
-
-const Footer = () => {
-  return (
-    <footer className={s.root}>
-      <div className={s.container}>
-        <div className={s.footerWrap}>
-          Coded with
-          <span className={s.heart}></span>
-          by You
-        </div>
-      </div>
-    </footer>
-  );
-};
+import Footer from "./Footer";
 
 export default Footer;
