+++ src/components/Heading/index.js
@@ -1,0 +1,3 @@
+import Heading from './Heading.js';
+
+export default Heading;
