+++ src/components/Heading/style.module.css
@@ -1,0 +1,5 @@
+.root {
+  color: #fff;
+  font-size: 54px;
+  margin-bottom: 48px;
+}
