+++ src/components/Slider/index.js
@@ -1,21 +1,3 @@
-import s from "./style.module.css";
-
-const Slider = () => {
-  return (
-    <section className={s.section}>
-      <div className={s.slider}>
-        <div className={`${s.container} ${s.sliderContent}`}>
-          <h1 className={s.header}>Wow</h1>
-
-          <h2 className={s.subheader}>Wow.Wow.Wow</h2>
-
-          <div className={s.call}>
-            <button className={s.button}>Wow</button>
-          </div>
-        </div>
-      </div>
-    </section>
-  );
-};
+import Slider from "./Slider";
 
 export default Slider;
