+++ src/components/Slider/style.module.css
